Keep VALUE instructions in a sorted order
Some passes, such as ccp, may change the order of VALUE instructions
(without changing the IR in other ways) by doing optimizations such
as moving a constant into a phi node. For example,

   <bb 2> :
   i_20 = 0;
   goto <bb 4>;

   <bb 4> :
   # i_18 = PHI <i_20(2), i_27(3)>
   ...

is changed to

   <bb 2> :
   goto <bb 4>;

   <bb 4> :
   # i_18 = PHI <0(2), i_27(3)>
   ...

We therefore keep the VALUE instruction sorted to avoid IR differences.

+++ lib/smtgcc.h
@@ -261,7 +261,7 @@ struct Function {
 public:
   std::string name;
   std::vector<Basic_block *> bbs;
-  std::map<std::pair<unsigned __int128, uint32_t>, Instruction *> values;
+  std::map<std::pair<uint32_t, unsigned __int128>, Instruction *> values;
   Instruction *last_value_inst = nullptr;
 
   Basic_block *build_bb();
